[chore] export App types
