+++ packages/kit/src/core/preview/index.js
@@ -34,7 +34,7 @@ export async function preview({
 
 	const app_file = resolve(cwd, `${SVELTE_KIT}/output/server/app.js`);
 
-	/** @type {import('types/internal').App} */
+	/** @type {import('types/app').App} */
 	const app = await import(pathToFileURL(app_file).href);
 
 	/** @type {import('sirv').RequestHandler} */
