+++ packages/kit/src/runtime/server/index.js
@@ -8,7 +8,7 @@ import { coalesce_to_error } from '../utils.js';
 import { hash } from '../hash.js';
 
 /**
- * @param {import('types/internal').Incoming} incoming
+ * @param {import('types/app').IncomingRequest} incoming
  * @param {import('types/internal').SSRRenderOptions} options
  * @param {import('types/internal').SSRRenderState} [state]
  */
