+++ packages/kit/types/index.d.ts
@@ -3,6 +3,7 @@
 
 import './ambient-modules';
 
+export { App, IncomingRequest } from './app';
 export { Adapter, AdapterUtils, Config, PrerenderErrorHandler, ValidatedConfig } from './config';
 export { EndpointOutput, RequestHandler } from './endpoint';
 export { ErrorLoad, ErrorLoadInput, Load, LoadInput, LoadOutput, Page } from './page';
